# src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "SpaceRouter Marketplace",
        "contact": {
            "name": "ESIEESPACE Network",
            "url": "http://esieespace.fr",
            "email": "contact@esieespace.fr"
        },
        "license": {
            "name": "GPL-3.0",
            "url": "https://github.com/SpaceRouter/authentication_server/blob/louis/LICENSE"
        },
        "version": "0.1"
    },
    "host": "localhost:8081",
    "basePath": "/",
    "paths": {
        "/v1/developer/{search}": {
            "get": {
                "description": "Get all stacks information uploaded by a user",
                "produces": [
                    "application/json"
                ],
                "summary": "Get stack by user ID",
                "operationId": "get_stack_by_user_ID",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/forms.StackSearchResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/forms.StackSearchResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/forms.StackSearchResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/forms.StackSearchResponse"
                        }
                    }
                }
            }
        },
        "/v1/search/stack/{search}": {
            "get": {
                "description": "Search for stacks reduced information by name",
                "produces": [
                    "application/json"
                ],
                "summary": "Search for stacks by name",
                "operationId": "get_stack_search",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/forms.StackSearchResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/forms.StackSearchResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/forms.StackSearchResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/forms.StackSearchResponse"
                        }
                    }
                }
            }
        },
        "/v1/stack/{id}": {
            "get": {
                "description": "Get all stack information from stack ID",
                "produces": [
                    "application/json"
                ],
                "summary": "Get stack by ID",
                "operationId": "get_stack_by_id",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/forms.StackResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/forms.StackResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/forms.StackResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/forms.StackResponse"
                        }
                    }
                }
            }
        },
        "/v1/stack_by_name/{name}": {
            "get": {
                "description": "Get all stack information from stack name",
                "produces": [
                    "application/json"
                ],
                "summary": "Get stack by Name",
                "operationId": "get_stack_by_name",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/forms.StackResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/forms.StackResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/forms.StackResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/forms.StackResponse"
                        }
                    }
                }
            }
        },
        "/v1/stacks": {
            "get": {
                "description": "Get all stacks",
                "produces": [
                    "application/json"
                ],
                "summary": "Get all stacks",
                "operationId": "get_all_stacks",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/forms.StackSearchResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/forms.StackSearchResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/forms.StackSearchResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/forms.StackSearchResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "forms.StackInfo": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "developer": {
                    "$ref": "#/definitions/models.Developer"
                },
                "icon": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "forms.StackResponse": {
            "type": "object",
            "properties": {
                "developer": {
                    "$ref": "#/definitions/models.Developer"
                },
                "message": {
                    "type": "string"
                },
                "ok": {
                    "type": "boolean"
                },
                "stack": {
                    "$ref": "#/definitions/models.Stack"
                }
            }
        },
        "forms.StackSearchResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "ok": {
                    "type": "boolean"
                },
                "stacks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/forms.StackInfo"
                    }
                }
            }
        },
        "models.Developer": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "website": {
                    "type": "string"
                }
            }
        },
        "models.EnvVar": {
            "type": "object",
            "properties": {
                "defaultValue": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "serviceID": {
                    "type": "integer"
                }
            }
        },
        "models.Network": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "serviceID": {
                    "type": "integer"
                }
            }
        },
        "models.NetworkDeclaration": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "stackID": {
                    "type": "integer"
                }
            }
        },
        "models.Port": {
            "type": "object",
            "properties": {
                "port": {
                    "type": "integer"
                },
                "serviceID": {
                    "type": "integer"
                }
            }
        },
        "models.Service": {
            "type": "object",
            "properties": {
                "envs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.EnvVar"
                    }
                },
                "image": {
                    "type": "string"
                },
                "imageVersion": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "networks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Network"
                    }
                },
                "ports": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Port"
                    }
                },
                "stackID": {
                    "type": "integer"
                },
                "volumes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Volume"
                    }
                }
            }
        },
        "models.Stack": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "developerID": {
                    "type": "integer"
                },
                "icon": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "networks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.NetworkDeclaration"
                    }
                },
                "services": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Service"
                    }
                },
                "volumes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.VolumeDeclaration"
                    }
                }
            }
        },
        "models.Volume": {
            "type": "object",
            "properties": {
                "mountPoint": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "serviceID": {
                    "type": "integer"
                }
            }
        },
        "models.VolumeDeclaration": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "stackID": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}